Add GameStatistics class and move methods to new class
