+++ spec/game_statistics_spec.rb
@@ -1,0 +1,9 @@
+require './lib/game_statistics'
+
+RSpec.describe GameStatistics do
+
+
+
+
+
+end
